Added tutorial

+++ Classes/GameController.cs
@@ -324,7 +324,7 @@ namespace SLIL.Classes
 
         internal void StopGame(int win) => Game.StopGame(win);
 
-        public void SetCustom(bool custom, int CustomWidth, int CustomHeight, string CustomMap, int customX, int customY) => Game.SetCustom(custom, CustomWidth, CustomHeight, CustomMap, customX, customY);
+        public void SetCustom(bool custom, int CustomWidth, int CustomHeight, string CustomMap, double customX, double customY) => Game.SetCustom(custom, CustomWidth, CustomHeight, CustomMap, customX, customY);
 
         public int GetPing() => peer.Ping;
